+++ CHANGELOG.md
@@ -12,6 +12,11 @@ Created By: BENGUERGOURA Oussama and Hadjersi Mohamed | 01/09/2024
 
 ### Added
 
+* Adding validation to `UserLoginDTO`.
+* Adding `CustomAuthenticationProvider` to implement retrieveUser method using `pesudo` and `companyCode`.
+* Adding `findByPseudoAndCompanyCode` in `CommonUserRepository`.
+* Adding `companyCode` to authentication method.
+* Adding `AuthService` to handler authentication and token generation.
 * Adding `NoArgsConstructor` to `CommonDto`.
 * Adding `id` to `CommonRoleDTO`.
 * Adding `role` and `email` to `CommonUserDTO`.
@@ -19,6 +24,7 @@ Created By: BENGUERGOURA Oussama and Hadjersi Mohamed | 01/09/2024
 ### Changed
 
 * Change `UserConstants` to `CommonUserConstants`.
+* Renaming `UserDto` to `UserResponse`.
 * Authentification checking now from database and not from InMemory saved users.
 * Renaming `FunctionalityMapper` and `ModuleMapper` on adding `Common` prefix